+++ bundlers/browser/BrowserMultiPack.ts
@@ -1,6 +1,5 @@
 ﻿import path = require("path");
 import stream = require("stream");
-import browserify = require("browserify");
 import CombineSourceMap = require("combine-source-map");
 import through2 = require("through2");
 import umd = require("umd");
@@ -92,23 +91,23 @@ module BrowserMultiPack {
     /** Override browserify's standard 'pack' pipeline step with a custom 'browser-pack' implementation that writes to multiple output bundles.
      * This requires overwriting browserif.prototype._createPipeline() and setting the 'bundleBldr' setBundleSourceCreator() callback
      * @param bundleBldr the bundle builder to modify
-     * @param _browserify the browserify instance to modify to output multiple bundle streams
+     * @param _bundler the browserify instance to modify to output multiple bundle streams
      * @param getMultiBundleOpts a function which returns a MultiBundleOptions object containing the options to build the bundle streams.
      * This function gets called when browserify.bundle() is called, which happens right at the beginning of BundleBuilder.compileBundle() (which calls BrowserifyHelper.setupRebundleListener())
      * @param getOpts options related to setting up the bundle streams
      */
-    export function overrideBrowserifyPack(
-        bundleBldr: BundleBuilder.Builder<browserify.BrowserifyObject>,
-        _browserify: { prototype: { _bpack: any; _createPipeline(opts: any): any } },
+    export function overrideBrowserifyPack<TBundler extends { bundle(): NodeJS.ReadableStream }>(
+        bundleBldr: BundleBuilder.Builder<TBundler>,
+        _bundler: { prototype: { _bpack: any; _createPipeline(opts: any): any } },
         getMultiBundleOpts: () => MultiBundleOptions
     ) {
-        var origCreatePipeline = _browserify.prototype["_createPipeline"];
+        var origCreatePipeline = _bundler.prototype["_createPipeline"];
         var newBpack: { baseStream: stream.Transform; bundleStreams: BundleStream<stream.Transform>[] };
         var updateDeps: string[] | null = null;
 
         // Override browserify._createPipeline() to replace the 'pack' pipeline step with a custom browser-pack implementation
         // gets called when browserify instance is created or when reset() or bundle() are called
-        _browserify.prototype["_createPipeline"] = function _createPipelineBundleSpliterCustomization(this: typeof _browserify["prototype"], createPipeOpts: any) {
+        _bundler.prototype["_createPipeline"] = function _createPipelineBundleSpliterCustomization(this: typeof _bundler["prototype"], createPipeOpts: any) {
             var pipeline = origCreatePipeline.call(this, createPipeOpts);
             var packPipe = pipeline.get("pack");
             var oldBpack = packPipe.pop();
@@ -123,9 +122,9 @@ module BrowserMultiPack {
         };
 
         // Consume the browserify bundle and return the multiple pack bundles
-        bundleBldr.setBundleSourceCreator(function multiBundleStreamCreator(browserify: browserify.BrowserifyObject, updateEvent?: { [key: string]: any }) {
+        bundleBldr.setBundleSourceCreator(function multiBundleStreamCreator(bundler: TBundler, updateEvent?: { [key: string]: any }) {
             if (updateEvent != null) { updateDeps = Object.keys(updateEvent).map((k) => updateEvent[k]); }
-            var brwsBundle = browserify.bundle();
+            var brwsBundle = bundler.bundle();
             var res: any[] = [];
             // When the bundle stream has available data, read it so that the stream ends
             brwsBundle.on("readable", () => {